Fix WireMock test utility and update repo URL
