+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
   - name: Request support
-    url: https://github.com/SaladTechnologies/saladcloud-job-queue-worker-sdk/discussions
+    url: https://github.com/SaladTechnologies/salad-cloud-job-queue-worker/discussions
     about: Please use GitHub discussions for general chat and community-provided support.
