+++ .github/SECURITY.md
@@ -8,7 +8,7 @@ _Please do not use public GitHub issues to report a security vulnerability._
 
 Instead, please do one of the following:
 
-- Open a GitHub security advisory on the [GitHub repository](https://github.com/SaladTechnologies/saladcloud-job-queue-worker-sdk/security/advisories/new)
+- Open a GitHub security advisory on the [GitHub repository](https://github.com/SaladTechnologies/salad-cloud-job-queue-worker/security/advisories/new)
 - Send an email to [`dev@salad.com`](mailto:dev@salad.com)
 
 Please include as much information as you can provide to help us better understand the nature and scope of the issue.
