+++ .github/SUPPORT.md
@@ -1,3 +1,3 @@
 # Support
 
-We use [GitHub discussions](https://github.com/SaladTechnologies/saladcloud-job-queue-worker-sdk/discussions) for general chat and community-provided support. The community often chimes in with helpful advice when you have a question, and you may also find yourself providing answers and helping others. Be sure to review the [code of conduct](./CODE_OF_CONDUCT.md) before participating.
+We use [GitHub discussions](https://github.com/SaladTechnologies/salad-cloud-job-queue-worker/discussions) for general chat and community-provided support. The community often chimes in with helpful advice when you have a question, and you may also find yourself providing answers and helping others. Be sure to review the [code of conduct](./CODE_OF_CONDUCT.md) before participating.
