+++ README.md
@@ -1,9 +1,9 @@
-# SaladCloud Job Queue Worker SDK
+# SaladCloud Job Queue Worker
 
-[![License](https://img.shields.io/github/license/SaladTechnologies/saladcloud-job-queue-worker-sdk)](./LICENSE)
+[![License](https://img.shields.io/github/license/SaladTechnologies/salad-cloud-job-queue-worker)](./LICENSE)
 
 <center>
-  <a href="https://github.com/SaladTechnologies/saladcloud-job-queue-worker-sdk"><img alt="SaladCloud Job Queues" src="./images/saladcloud-job-queues-banner.png" width="100%" /></a>
+  <a href="https://github.com/SaladTechnologies/salad-cloud-job-queue-worker"><img alt="SaladCloud Job Queues" src="./images/saladcloud-job-queues-banner.png" width="100%" /></a>
 </center>
 
-This project contains the SaladCloud Job Queue Worker SDK and samples.
+This project contains the SaladCloud Job Queue Worker, SDK, and samples. Refer to the [Job Queues documentation](https://docs.salad.com/container-engine/job-queues) for more information on using this with your SaladCloud-deployed workloads.
