+++ internal/workers/workers.go
@@ -11,11 +11,11 @@ import (
 	"time"
 
 	"github.com/cloudflare/backoff"
-	metadata_api "github.com/saladtechnologies/saladcloud-job-queue-worker-sdk/internal/apis/metadata"
-	queues_api "github.com/saladtechnologies/saladcloud-job-queue-worker-sdk/internal/apis/queues"
-	"github.com/saladtechnologies/saladcloud-job-queue-worker-sdk/pkg/config"
-	"github.com/saladtechnologies/saladcloud-job-queue-worker-sdk/pkg/jobs"
-	"github.com/saladtechnologies/saladcloud-job-queue-worker-sdk/pkg/log"
+	metadata_api "github.com/saladtechnologies/salad-cloud-job-queue-worker/internal/apis/metadata"
+	queues_api "github.com/saladtechnologies/salad-cloud-job-queue-worker/internal/apis/queues"
+	"github.com/saladtechnologies/salad-cloud-job-queue-worker/pkg/config"
+	"github.com/saladtechnologies/salad-cloud-job-queue-worker/pkg/jobs"
+	"github.com/saladtechnologies/salad-cloud-job-queue-worker/pkg/log"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
